rename generator-base-blueprint to generator
