+++ generators/base/index.mts
@@ -17,5 +17,5 @@
  * limitations under the License.
  */
 
-export { default } from './generator-base-blueprint.mjs';
+export { default } from './generator.mjs';
 export { default as command } from './command.mjs';
